+++ noodles-bgzf/src/lib.rs
@@ -42,17 +42,19 @@ pub(crate) mod deflate;
 mod gz;
 pub mod gzi;
 pub mod io;
-mod multithreaded_reader;
 pub mod multithreaded_writer;
 pub mod virtual_position;
 
-pub use self::{
-    multithreaded_reader::MultithreadedReader, multithreaded_writer::MultithreadedWriter,
-    virtual_position::VirtualPosition,
-};
+pub use self::{multithreaded_writer::MultithreadedWriter, virtual_position::VirtualPosition};
 
 #[deprecated(since = "0.38.0", note = "Use `bgzf::io::IndexedReader` instead.")]
 pub use self::io::IndexedReader;
+
+#[deprecated(
+    since = "0.38.0",
+    note = "Use `bgzf::io::MultithreadedReader` instead."
+)]
+pub use self::io::MultithreadedReader;
 
 #[deprecated(since = "0.38.0", note = "Use `bgzf::io::Reader` instead.")]
 pub use self::io::Reader;