+++ noodles-bgzf/CHANGELOG.md
@@ -4,12 +4,13 @@
 
 ### Changed
 
-  * bgzf: Move readers (`IndexedReader` and `Reader`) and writer (`Writer`) to
-    `io` module.
+  * bgzf: Move readers (`IndexedReader`, `MultithreadedReader`, and `Reader`)
+    and writer (`Writer`) to `io` module.
 
 ### Deprecated
 
-  * bgzf: Deprecate `IndexedReader`, `Reader` and `Writer`.
+  * bgzf: Deprecate `IndexedReader`, `MultithreadedReader`, `Reader` and
+    `Writer`.
 
     These moved to the `io` module.